requirements.txt
django-model-search>=2.1
django-modeltranslation>=0.14.4